Add Go bundler for genezio-runtime
Implement bundler and handler generator for genezio runtime.
The go code is compiled as a plugin which can be loaded at
runtime in another Go application.
The plugin exports a function to handle requests, which
is called by the runtime go wrapper after loading the plugin.
