+++ src/bundlers/go/lambdaGoMain.ts
@@ -9,12 +9,11 @@ package main
 import (
 	"encoding/json"
     "encoding/base64"
-	"net/http"
-	"os"
-    "io"
     "errors"
+    "context"
     "path"
     "github.com/Genez-io/genezio_types"
+    "github.com/aws/aws-lambda-go/lambda"
 
     {{#imports}}
     {{#named}}{{name}} {{/named}}"{{{path}}}"
@@ -81,12 +80,11 @@ const (
     JsonRpcMethod MethodType = "jsonrpc"
 )
 
-func sendError(w http.ResponseWriter, err error, methodType MethodType) {
+func sendError(err error, methodType MethodType) *Response {
     genezioError := make(map[string]interface{})
     byteError, error := json.Marshal(err)
     if error != nil {
-        http.Error(w, error.Error(), http.StatusInternalServerError)
-        return
+        return nil
     }
     json.Unmarshal(byteError, &genezioError)
 	var responseError ResponseBodyError
@@ -104,10 +102,9 @@ func sendError(w http.ResponseWriter, err error, methodType MethodType) {
     }
     responseErrorByte, err := json.Marshal(responseError)
     if err != nil {
-        http.Error(w, err.Error(), http.StatusInternalServerError)
-        return
+        return nil
     }
-	response := Response{
+	response := &Response{
 		StatusCode: "200",
 		Body:       string(responseErrorByte),
 		Headers: map[string]string{
@@ -115,29 +112,14 @@ func sendError(w http.ResponseWriter, err error, methodType MethodType) {
 			"X-Powered-By": "genezio",
 		},
 	}
-	responseByte, err := json.Marshal(response)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	w.Header().Set("Content-Type", "application/json")
-	io.WriteString(w, string(responseByte))
+    return response
 }
 
-func handleRequest(w http.ResponseWriter, r *http.Request) {
-	var event Event
+func handleRequest(context context.Context, event *Event) (*Response, error) {
 	var body EventBody
 	var responseBody ResponseBody
 
-	err := json.NewDecoder(r.Body).Decode(&event)
-	if err != nil {
-        sendError(w, err, JsonRpcMethod)
-		return
-	}
-
-    defer r.Body.Close()
-
-    class := {{class.packageName}}.New()
+	class := {{class.packageName}}.New()
 
     var isJsonRpcRequest bool
 
@@ -155,13 +137,13 @@ func handleRequest(w http.ResponseWriter, r *http.Request) {
         case "{{name}}":
             err := class.{{name}}()
             if err != nil {
-                sendError(w, err, CronMethod)
-                return
+                errorResponse := sendError(err, CronMethod)
+                return errorResponse, nil
             }
         {{/cronMethods}}
         default:
-            sendError(w, errors.New("cron method not found"), CronMethod)
-            return
+            errorResponse := sendError(errors.New("Cron method not found"), CronMethod)
+            return errorResponse, nil
         }
     } else if !isJsonRpcRequest {
         genezioRequest := genezio_types.GenezioHttpRequest{
@@ -174,8 +156,8 @@ func handleRequest(w http.ResponseWriter, r *http.Request) {
         if event.IsBase64Encoded {
             bodyDecoded, err := base64.StdEncoding.DecodeString(event.Body)
             if err != nil {
-                sendError(w, err, HttpMethod)
-                return
+                errorResponse := sendError(err, HttpMethod)
+                return errorResponse, nil
             }
             genezioRequest.Body = string(bodyDecoded)
         } else {
@@ -195,31 +177,37 @@ func handleRequest(w http.ResponseWriter, r *http.Request) {
         case "{{name}}":
             result, err = class.{{name}}(genezioRequest)
             if err != nil {
-                sendError(w, err, HttpMethod)
-                return
+                errorResponse := sendError(err, HttpMethod)
+                return errorResponse, nil
             }
         {{/httpMethods}}
         default:
-            sendError(w, errors.New("http method not found"), HttpMethod)
-            return
+            errorResponse := sendError(errors.New("Http method not found"), HttpMethod)
+            return errorResponse, nil
         }
-
-        resultBody, err := json.Marshal(result.Body)
+        responseBody, err := json.Marshal(result.Body)
         if err != nil {
-            sendError(w, err, HttpMethod)
-            return
+            errorResponse := sendError(err, HttpMethod)
+            return errorResponse, nil
         }
-        result.Body = string(resultBody)
-        response, err := json.Marshal(result)
-        if err != nil {
-            sendError(w, err, HttpMethod)
-            return
-        }  
-        w.Header().Set("Content-Type", "application/json")
-        w.WriteHeader(http.StatusOK)
-        io.WriteString(w, string(response))
-        return
+        var responseHeaders map[string]string
+        if result.Headers != nil {
+            responseHeaders = *result.Headers
+        }
+        response := &Response{
+            StatusCode: result.StatusCode,
+            Body:       string(responseBody),
+            Headers:    responseHeaders,
+        }
+        return response, nil
     } else {
+        eventBody := []byte(event.Body)
+        // Decode the request body into struct and check for errors
+        err := json.Unmarshal(eventBody, &body)
+        if err != nil {
+            errorResponse := sendError(err, JsonRpcMethod)
+            return errorResponse, nil
+        }
         // Call the appropriate method
         switch body.Method {
         {{#jsonRpcMethods}}
@@ -229,8 +217,8 @@ func handleRequest(w http.ResponseWriter, r *http.Request) {
             {{/parameters}}
             {{^isVoid}}result, {{/isVoid}}err {{^isVoid}}:{{/isVoid}}= class.{{name}}({{#parameters}}param{{index}}{{^last}}, {{/last}}{{/parameters}})
             if err != nil {
-                sendError(w, err, JsonRpcMethod)
-                return
+                errorResponse := sendError(err, JsonRpcMethod)
+                return errorResponse, nil
             }
             {{^isVoid}}
             responseBody.Result = result
@@ -240,25 +228,26 @@ func handleRequest(w http.ResponseWriter, r *http.Request) {
         case "{{class.name}}.{{name}}":
             err := class.{{name}}()
             if err != nil {
-                sendError(w, err, JsonRpcMethod)
-                return
+                errorResponse := sendError(err, JsonRpcMethod)
+                return errorResponse, nil
             }
         {{/cronMethods}}
         default:
-            sendError(w, errors.New("method not found"), JsonRpcMethod)
-            return
+            errorResponse := sendError(errors.New("Method not found"), JsonRpcMethod)
+            return errorResponse, nil
         }
-        responseBody.Id = body.Id
-        responseBody.Jsonrpc = body.Jsonrpc
     }
+
+	responseBody.Id = body.Id
+	responseBody.Jsonrpc = body.Jsonrpc
 
     bodyString, err := json.Marshal(responseBody)
     if err != nil {
-        sendError(w, err, JsonRpcMethod)
-        return
+        errorResponse := sendError(err, JsonRpcMethod)
+        return errorResponse, nil
     }
 
-	response := Response{
+	response := &Response{
         StatusCode: "200",
 		Body:       string(bodyString),
 		Headers: map[string]string{
@@ -266,20 +255,10 @@ func handleRequest(w http.ResponseWriter, r *http.Request) {
 		},
 	}
 
-    // Encode the struct into JSON and check for errors
-	responseByte, err := json.Marshal(response)
-	if err != nil {
-        sendError(w, err, JsonRpcMethod)
-		return
-	}
-	w.Header().Set("Content-Type", "application/json")
-    w.WriteHeader(http.StatusOK)
-	io.WriteString(w, string(responseByte))
+    return response, nil
 }
 
 func main() {
-	port := os.Args[1]
-	http.HandleFunc("/", handleRequest)
-	http.ListenAndServe(":"+port, nil)
+    lambda.Start(handleRequest)
 }
 `;
