+++ src/bundlers/utils.ts
@@ -9,7 +9,7 @@ import { NodeJsBinaryDependenciesBundler } from "./node/nodeJsBinaryDependencies
 import { BundlerComposer } from "./bundlerComposer.js";
 import { DartBundler } from "./dart/dartBundler.js";
 import { KotlinBundler } from "./kotlin/kotlinBundler.js";
-import { GoBundler } from "./go/goBundler.js";
+import { NewGoBundler } from "./go/goBundler.js";
 import { debugLogger, printAdaptiveLog } from "../utils/logging.js";
 import { createTemporaryFolder } from "../utils/file.js";
 import { ProjectConfiguration } from "../models/projectConfiguration.js";
@@ -59,7 +59,7 @@ export async function bundle(
             break;
         }
         case "go": {
-            bundler = new GoBundler();
+            bundler = NewGoBundler(projectConfiguration);
             break;
         }
         default:
